Moved the generation of the julia API into a different script
